OAK-857 MongoMK: support for many child nodes (bugfix)
git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1491397 13f79535-47bb-0310-9956-ffa450edef68

+++ oak-mongomk/src/main/java/org/apache/jackrabbit/mongomk/MongoMK.java
@@ -956,7 +956,7 @@ public class MongoMK implements MicroKernel {
         }
         if (c.hasMore) {
             // TODO use a better way to notify there are more children
-            json.key(":childNodeCount").value(Integer.MAX_VALUE);
+            json.key(":childNodeCount").value(Long.MAX_VALUE);
         } else {
             json.key(":childNodeCount").value(c.children.size());
         }
